requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-environ==0.11.2  # ileride Postgres'e geçiş için
gunicorn==22.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0
django-tenants
python-dotenv
psycopg2-binary>=2.9
setuptools>=65.5.1
Pillow==10.4.0